Kwoka_Moscicka_Wach/requirements.txt
transformers
torch
accelerate
huggingface_hub
einops
pillow
pyyaml
torchvision
bitsandbytes
